pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dev-pilot"
version = "0.2.0"
requires-python = ">=3.10"
dependencies = ["pyyaml>=6.0", "click>=8.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "ruff>=0.4.0"]

[project.scripts]
dev-pilot = "core.cli:main"

[tool.setuptools.packages.find]
include = ["core*"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
